refactor(tests): decompose matrix_utils tests into connect-domain modules
- delete legacy tests/test_matrix_utils_auth.py to avoid duplication
- add new tests for connect behavior:
- test_matrix_utils_connect.py
- test_matrix_utils_connect_credentials.py
- test_matrix_utils_connect_e2ee.py
- test_matrix_utils_connect_rooms.py
- test_matrix_utils_connect_sync.py
- update docs/dev/TESTING_GUIDE.md to reflect the new testing structure and removal of the monolith
- ensure tests reference the new split layout and maintain compatibility
